roachtest: eliminate the concept of subtests
The point of subtests was that all the subtests under a root share a
cluster. This patch introduces more flexible policies for sharing
clusters across tests (only the declaration of these policies is here;
the implementation is in future commits). The policies replace the need
for subtests, so they go away, simplifying the code considerably.

Release note: None