storage,sql: add a fast mode to CheckConsistency and expose via SQL
It has always been difficult to diagnose consistency check failures for
to three reasons:

1. they're rare (and often go undetected)
2. they require the data directories to meaningfully investigate
3. failures often occur long after the likely triggers

This PR works to address some of all of these. Note that none of this is
for end user consumption at this point.

We add a "quick" mode to CheckConsistency. In that mode, the hash of a
replica is simply the hash of its RangeAppliedState. We know only of one
incident in which a detected replica inconsistency did not include this
key. This is because the consistently replicated MVCCStats are stored
under this key, and because consistency failures are almost always
caused by to an "atomic" failure to apply a batch.
That is, when a write gets lost, it's very likely that on the replica
losing the write the corresponding stats update will also be missing.

The only known exception is one recent incident involving only time
series data, where there is vague hope that the problem might instead
have been caused by some problem specific to RocksDB merges.

Having the quick mode available makes it feasible to run the consistency
checks throughout our various roachtests, and to detect more
inconsistencies close to when they happen, both of which will naturally
improve our ability to get to the bottom of them.

Consistency checks are usually run on a queue at low frequency, making
it hard to predict when a given range will be checked. To make this more
convenient, we add a sql builtin `crdb_internal.check_consistency` which
does not trigger fatal node errors when it detects an inconsistency, and
takes three parameters:

1. whether to use the stats only (bool), enabling the quick mode
2. start key (byte) (defaults to min possible)
3. end key (byte)   (defaults to max possible)

These three inputs roughly translate to a

    CheckConsistencyRequest{Key: key, EndKey: end_key, Mode: <mode>}.

and the output is returned as a array of tuples, which can be consumed
conveniently via the desired variant of

```
SELECT (t).* FROM unnest(crdb_internal.check_consistency(true, '\x02', '\x04')) as t
WHERE (t).status != 'RANGE_CONSISTENT';

  range_id | status | detail
+----------+--------+--------+
(0 rows)
```

Release note: None