rpc: add grpc interceptors for clock signals
Before this patch, exchanging of clock signals between nodes was done
only through the Batch RPCs, where the client->server timestamp was
ba.Timestamp, and the server->client timestamp was res.Timestamp or
pErr.Timestamp.
This code was quite unprincipled. First of all there are other ways that
nodes communicate besides Batch RPCs. Second of all, ba.Timestamp is not
the client's timestamp. It is generally a transaction's timestamp, which
is generally older.

This patch replaces the existing mechanism with one based on gRPC
interceptors. A pair of client/server interceptors pass timestamps back
and forth on every RPC. This is supposed to result in better
synchronized clocks which is good for performance (because sync clocks
=> fewer uncertainty restarts) but also beyond performance there are
cases where we rely on nodes having bumped their timestamps at key
moments and currently it's not very clear that that happens (or at the
very least this change will put our footing on firmer ground because
critical clock synchronization is less likely to silently dissapear with
refactorings).

Touches #7526

TODO: migration story
TODO: interceptors for streaming RPCs.

Release note: None