storage: remove dependency from replication queue to the split queue
Before this patch, the replication queue would refuse to process ranges
that it thought needed splits because of their size. The idea was to let
the split queue process them first. There's no reason for this,
particularly since we've introduced the write-backpressuring mechanism
which keeps ranges from getting too large (although not all requests are
backpressured, so theoretically ranges could still grow larger than we'd
like).
The check had not considered unsplittable ranges (i.e. ranges with only
(versions of) one row or some system ranges). These can't be split but
the replication queue was disconsidering them none the less.

This change was motivated by the unsplittable SystemConfigSpan range.
This kind of system range is not backpressured, but we still don't want
the replication queue to dismiss it when it's too large (since, again,
it can't be split).

Release note (bug fix): Ranges consisting of only one row (and
historical versions of that row) are now correctly up-replicating.