config,sqlbase: move computation of splits for SQL tables to sqlbase
Before this patch, the config package was in charge of figuring out the
split points to be created by the split queue - generally between tables
and between partitions where there's an applicable zone config.
This patch moves the code for determining the splits to sqlbase, where
it belongs better: there was no good reason for the config package to
deal with table keys and, in the future we'll likely tie zone configs
and table descriptors more and more (and the config package doesn't have
access to table descriptor). Besides generally seeming like a
good direction, I've got some more immediate optimizations that I want
to make to the splitting code that requires access to table descriptor
when deciding the split points.

This patch also rationalizes the remaining code dealing with splits in
the config package, which was flirting with unreadability.

Release note: None