sql: track DDL executed in current txn
This patch changes the meaning of connEx.extraTxnState.numDDL to be the
number of DDL statements executed in the current transaction, not on the
respective connection since the connection was established. Fields in
extraTxnState are supposed to be tied to the current transaction, and
also I need the new semantics for the next patch since restarting
transaction that ran DDL statements needs special handling.

For the purposes for this field was used (i.e. figuring out which
savepoints we can rollback to), both definitions work the same.

Release note: None