kvserver: hookup the new closedts tracker to outgoing proposals
This patch adds a closed timestamp field to every Raft command. These
timestamps are assigned at the moment when a proposal is flushed out of
the proposal buffer.
In order to not close timestamps below any currently-evaluating
requests, the recent request tracker is hooked up to the request write
path.

Nobody does anything with the closed timestamps carried by Raft commands
yet; that part is next.

Release note: None