roachtest: run against a set of existing clusters
Before this patch, you could specify one cluster with `roachprod run
--cluster=foo --parallelism=1`, and tests would run one at a time
against that one cluster. All the tests needed to be compatible with the
spec of the existing cluster.

This patch adds --existing-clusters-name=foo-
--existing-clusters-count=n. Clusters foo-[1..n] are expected to exist.
Tests will run against these clusters.  There's also a new
--wipe-before-reuse[=false], which can be used to inhibit the wiping of
clusters in between passing tests.  The idea is to pre-create a bunch of
clusters, perhaps load them with data once, and then run tests against
them repeatedly.

 --parallelism is supported; it gets effectively truncated it it's > n.
The clusters can have different specs, and each test will find a cluster
that matches what the test wants. The runner will fail if there's no
cluster available for a test. The clusters are not destroyed when tests
fail or when the test runner exits. If combined with --debug, clusters
are not reused after a failed test.

The patch also adds a generic --test-flags that can be used to pass
options to tests. The tpccbench tests are taught to recognize the
"skip-cluster-create" option and skip some initialization, such that it
runs on a cluster that was already initialized.

With all this I can run tpccbench over and over on a pool of existing,
pre-loaded clusters without having each test do the slow loading of
data.

Release note: None