contextutil: add a Context with richer cancellation
When a context gets canceled, by contract of `context.Context` we can
only ever return the opaque errors `context.{Canceled,DeadlineExceeded}`
from `ctx.Err()`. This is not helpful and we often struggle with
locating the source of a context cancellation.

With this commit, you can do the following:

```go
ctx, cancel := contextutil.WithErrCancel(ctx)
go func() {
  cancel(errors.New("something bad happened in a random place")) // line X
}()
<-ctx.Done()
err := contextutil.ExtErr(ctx) // something bad happened in a random place
_, line, _, ok := errors.GetOneLineSource(err) // line X, true
}
```

This doesn't help if the context is canceled by
`context.With{Deadline,Timeout}` but we already have
`contextutil.RunWithTimeout` for that case.

Release note: None