democluster: don't intermix tracers when starting servers
In a demo cluster, starting each server was done through the cluster's
stopper, so each Server.Start(ctx) was running in a span created using
the cluster's Tracer. Each Server, though, has its own Tracer that it's
using to create child spans. This kind of mixing of Tracers is not going
to be supported any more, since it's fairly broken. As far as our own
Tracer is concerned, tracing across Tracers is not supported.

This patch fixes it by not propagating the cluster's span into Servers.

Release note: None