tracing: make a test robust against tracing being on
This patch makes the Tracer's behavior more configurable. Before, the
Tracer would generally not create spans when there was no explicit
reason to do so. If tracing was not enabled one way or another, it would
return the no-op span. This patch allows the Tracer to be configured to
always create spans - for example for the benefit of the active spans
registry. The respective option is not currently used, but will be soon.
The options for the Tracer are future-proof, also allowing tracing to be
enabled in tests, and also allowing tracing to be explicitly disabled
when the environment (e.g. tests) generally want it enabled. This last
capability is used by a logic test which, as written, does not like it
when tracing is turned on.

Release note: None