tracing: protect against use-after-finish when adding child spans
This patch adds protection for creating a child span in a finished
parent. This is illegal, as is any use of a span after Finish(). Still,
almost every other use of a span has protection against
use-after-Finish that makes the respective access behave reasonably.
This protection will also act as a point where I'll add an assertion, at
least in tests, verifying the absence of such illegal uses. So, this
patch adds similar protection to creating a child in a Finish()ed span.
In particular, it makes the new would-be child span act as a root and be
registered in the active spans registry. Before, new span would be added
to a finished parent, and so it would essentially fall by the wayside
and not make it to the registry.

Release note: None