tracing,server,distsql: optimize RPC trace creating for common RPCs
Before this patch, tracing across RPC boundaries is implemented by gRPC
interceptors that marshall the trace info from the client to the server
as gRPC metadata, and transparently open client and server spans.
Unfortunately, sending this metadata around the actual request is quite
expensive: it takes 10 allocations per RPC on the client and some on the
server. Some of these allocations can be optimized, but not all due to
the gRPC interfaces; this metadata is just not thought out to be
efficient. Additionally, back in the day, this metadata was measured to
produce a very significant increase in network packets (#17177). At the
time, this was reason enough to avoid this producing this metadata when
tracing is not enabled. I have not measured this again, nor was this
effect measured beyond the network packets metric, but I hope to enable
tracing enough such that, unless we do something else, this metadata
would again be always produced.

This patch optimizes tracing for the 3 most common RPCs in the system:
- Internal.Batch
- DistSQL.SetupFlow, DistSQL.FlowStream.

In the case of FlowStream, we simply do away with tracing. FlowStream is
a streaming RPC used by an Outbox to send its data to another node. The
outbox has its own span; the RPC's span did not have any information.
In fact I'm fairly sure we were not even collecting the RPC's stream in
the trace recordings.

For the other two, the respective clients and servers take on the
responsibility of marshalling the tracing information and creating
server-side spans. The client's trace info gets put into the request
protos - which dirties the respective protos with this concern. In the
case of Batch, the code and the protos were already concerned with
tracing because they dealt with returning the recording; that part was
never abstracted inside the interceptors.

BenchmarkTracing shows a significant reduction in allocations, and also
latency, for the 3-node tracing-on cases:

name                               old time/op    new time/op    delta
Tracing/3node/scan/trace=on-24        883µs ± 2%     847µs ± 1%  -4.11%  (p=0.008 n=5+5)
Tracing/3node/insert/trace=on-24     1.08ms ± 3%    1.04ms ± 3%  -3.94%  (p=0.016 n=5+5)

name                               old allocs/op  new allocs/op  delta
Tracing/3node/scan/trace=on-24          868 ± 0%       840 ± 0%  -3.20%  (p=0.008 n=5+5)
Tracing/3node/insert/trace=on-24        803 ± 0%       788 ± 0%  -1.89%  (p=0.008 n=5+5)

Release note: None