sql: don't create a tracing span for every pgwire cmd
Before this patch, processing each pgwire cmd was getting its own span
(when tracing is enabled). This was too much cost for too little
benefit. Generally, executing a query takes a few commands:
bind,describe,exec,sync. If the query is executed as an implicit txn,
the "exec" command is generally twice (once superficially in the
"no-txn" state and once in the "open" state). For simple queries, these
spans end up being about half of the total spans created for the query.
I've never looked at a span for anything other than the "exec", and
there only at the one in the "open" state. So this patch gets rid of
creating spans mechanically for every pgwire command. Instead, we create
spans only for preparing statements, and for executing a query in the
"open" state.

Release note: None