tracing: improve checking for consistent Tracer in children
Before this patch, attempting to create a child with a different Tracer
than the parent would panic is the parent was a "real span", but it
wouldn't if the parent was a no-op. This happenstence was allowing tests
to get away with mixing Tracers, as tests always run with tracing turned
off. Now that all the tests that were mixing Tracers were hopefully
fixed, this patch makes it illegal (panic) to mix Tracers even when
tracingt is not enabled, for consistency of behavior.

Release note: None