kvserver/concurrency: account for waitElsewhere time in contention events
Before this patch, no contention event was produced when requests were
waiting for locks by pushing the txn that currently has a conflicting
lock (or a reservation on a conflicting lock). This corresponds to the
waitElsewhere signal from the lock table. (waitElsewhere is not the
common case; the common case is for a request to wait in the lock table
(through the waitFor signal). waitElsewhere happens instead of waitFor
when the lock table is under memory pressure and sheds its state.)

This patch makes the waitElsewhere state be handled like the other
waiting states for the purposes of generating contention events.

Release note: None