partitionccl,testutils: make partitionccl tests work with a tenant
This patch makes all tests under partitionccl work when
serverutils.StartServer() decides to start a tenant inside the server
that it returns. StartServer does that randomly when
IsEnterpriseEnabled()==true (which it does because the main_test.go in
this pkg calls TestingEnableEnterprise).
Before this patch, when the random coin decided to add a tenant, the
tests ended up being silently skipped [1], because the binary with the
tests does not include all of the ccl packages. This skipping is bad.

The patch inhibits the creation of the tenant explicitly for a
few tests (where syntax unsupported in multi-tenancy is used - like zone
config stuff), and makes others more resilient.

The tests will continue to be skipped according to the coin like they
were before; this patch does not make them import all of ccl (but if
they did import all of ccl, they would now pass instead of fail). All of
ccl will likely be linked in the future by these tests (I got here by
working on such a change), but I'll do that separately.

Many other ccl subpackages probably need similar treatment; this pkg was
chosen arbitrarily to start with and see how it goes.

Tested with
dt pkg/ccl/partitionccl --stress -- --test_env="COCKROACH_TEST_TENANT_MODE=forceTenant"
and with adding
import _ "github.com/cockroachdb/cockroach/pkg/ccl"
to a file in the `partitionccl_test` pkg.

[1] https://github.com/cockroachdb/cockroach/blob/93d78cbcb5b196bdee7a71d3a5a8ce7774b751e6/pkg/testutils/serverutils/test_server_shim.go#L80

Release note: None
Epic: None