changefeedccl/kvfeed: make tests compatible with multi-tenancy
When multi-tenancy was enabled, these tests were executing SQL in one
tenant, but then attempting to login into the system tenant. This
problem was hidden because the tests are skipped dynamically when
multi-tenancy is enabled. This patch does not change the skipping, but
that will be done later when I start importing "ccl" from all the ccl
tests.

The patch makes the test work with a tenant, though, by not hardcoding
the use of the system codec.

Release note: None
Epic: None