testutils: capture Side-Eye snapshots in SucceedsSoon
Make testutils.SucceedsSoon() capture a snapshot of the test process
with Side-Eye when the deadline expires. The idea is to help with
debugging flaky unit tests: the snapshot should help us figure out where
the operation that the test was waiting on has gotten stuck.

The snapshot is generated if the SIDE_EYE_TOKEN environment variable is
set (it should be set to our organizations API token). The test log will
have a link to app.side-eye.io for visualizing the collected snapshot.

In the future, perhaps we can use Side-Eye to also collect execution
traces and CPU profiles for even more information.